moves copy functions into a module
